bundle update

+++ twidere/src/main/java/edu/tsinghua/spice/Task/SpiceAsyUploadTask.java
@@ -8,9 +8,12 @@ import android.util.Log;
 import org.mariotaku.restfu.annotation.method.POST;
 import org.mariotaku.restfu.http.RestHttpClient;
 import org.mariotaku.restfu.http.RestHttpRequest;
+import org.mariotaku.restfu.http.RestHttpResponse;
 import org.mariotaku.restfu.http.mime.FileTypedData;
 import org.mariotaku.restfu.http.mime.MultipartTypedBody;
 import org.mariotaku.twidere.BuildConfig;
+import org.mariotaku.twidere.Constants;
+import org.mariotaku.twidere.model.RequestType;
 import org.mariotaku.twidere.util.TwitterAPIFactory;
 
 import java.io.File;
@@ -20,19 +23,18 @@ import java.io.IOException;
 
 import edu.tsinghua.spice.Utilies.SpiceProfilingUtil;
 
-import static org.mariotaku.twidere.TwidereConstants.LOGTAG;
 import static org.mariotaku.twidere.util.Utils.copyStream;
 
 /**
  * Created by Denny C. Ng on 2/20/15.
  */
 
-public class SpiceAsyUploadTask extends AsyncTask<Object, Object, Object> {
+public class SpiceAsyUploadTask extends AsyncTask<Object, Object, Object> implements Constants {
 
-    private static final String PROFILE_SERVER_URL = "http://twidere-spice.mariotaku.org:18080/spice/usage";
+    public static final long UPLOAD_INTERVAL_MILLIS = 1000 * 60 * 60 * 24;
+    public static final String LAST_UPLOAD_TIME = "last_upload_time";
 
-    private static final String LAST_UPLOAD_DATE = "last_upload_time";
-    private static final double MILLSECS_HALF_DAY = 1000 * 60 * 60 * 12;
+    private static final String PROFILE_SERVER_URL = "http://spice.hot-mobile.org/spice/usage";
 
     private final Context context;
     private final RestHttpClient client;
@@ -65,9 +67,13 @@ public class SpiceAsyUploadTask extends AsyncTask<Object, Object, Object> {
             final MultipartTypedBody body = new MultipartTypedBody();
             body.add("file", new FileTypedData(tmp));
             builder.body(body);
-            client.execute(builder.build());
-            SpiceProfilingUtil.log("server has already received file " + tmp.getName());
-            tmp.delete();
+            builder.extra(RequestType.USAGE_STATISTICS);
+            final RestHttpResponse response = client.execute(builder.build());
+            if (response.isSuccessful()) {
+                SpiceProfilingUtil.log("server has already received file " + tmp.getName());
+                tmp.delete();
+            }
+            throw new IOException("Unable to send file");
         } catch (Exception e) {
             if (BuildConfig.DEBUG) {
                 Log.w(LOGTAG, e);
@@ -84,9 +90,9 @@ public class SpiceAsyUploadTask extends AsyncTask<Object, Object, Object> {
 
         final SharedPreferences prefs = context.getSharedPreferences("spice_data_profiling", Context.MODE_PRIVATE);
 
-        if (prefs.contains(LAST_UPLOAD_DATE)) {
-            final long lastUpload = prefs.getLong(LAST_UPLOAD_DATE, System.currentTimeMillis());
-            final double deltaDays = (System.currentTimeMillis() - lastUpload) / (MILLSECS_HALF_DAY * 2);
+        if (prefs.contains(LAST_UPLOAD_TIME)) {
+            final long lastUpload = prefs.getLong(LAST_UPLOAD_TIME, System.currentTimeMillis());
+            final double deltaDays = (System.currentTimeMillis() - lastUpload) / UPLOAD_INTERVAL_MILLIS;
             if (deltaDays < 1) {
                 SpiceProfilingUtil.log("Last uploaded was conducted in 1 day ago.");
                 return null;
@@ -96,7 +102,7 @@ public class SpiceAsyUploadTask extends AsyncTask<Object, Object, Object> {
         final File root = context.getFilesDir();
         final File[] spiceFiles = root.listFiles(new SpiceFileFilter());
         uploadToServer(spiceFiles);
-        prefs.edit().putLong(LAST_UPLOAD_DATE, System.currentTimeMillis()).apply();
+        prefs.edit().putLong(LAST_UPLOAD_TIME, System.currentTimeMillis()).apply();
         return null;
     }
 
@@ -140,5 +146,10 @@ public class SpiceAsyUploadTask extends AsyncTask<Object, Object, Object> {
                 SpiceProfilingUtil.log("put profile back failed");
             }
         }
+    }
+
+    public static long getLastUploadTime(final Context context) {
+        final SharedPreferences prefs = context.getSharedPreferences("spice_data_profiling", Context.MODE_PRIVATE);
+        return prefs.getLong(LAST_UPLOAD_TIME, -1);
     }
 }
